Update surface of Page, PagedListResponse and FixedSizeCollection
